pyproject.toml
[project]
name = "slack-ring-bot"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "ring-doorbell==0.9.13",
    "slack-sdk==3.38.0",
    "python-dotenv==1.0.0",
    "aiohttp==3.13.2",
]

[project.optional-dependencies]
check = ["ruff==0.4.0", "pre-commit==3.7.0"]

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
